requirements.txt
flask
flask-smorest
python-dotenv
flask-sqlalchemy
marshmallow
flask-jwt-extended
passlib
flask-migrate